move the Draco option tests to their own test_draco.py
They were in test_server.py only because that suite runs on lean wheels
(built without remote-access), where test_remote_access.py is skipped
but the DracoEncodeOptions/DracoMethod classes are still registered. A
WebSocket-server suite is the wrong home for them; give them a dedicated
always-run file whose module docstring records why it isn't gated on
remote-access. No test logic changes.
